server: split connection related decls into their own header file
