+++ server/server.h
@@ -23,31 +23,11 @@
 
 #include <stdint.h>
 #include "twopence.h"
-#include "protocol.h"
-#include "socket.h"
-#include "transaction.h"
-
-typedef struct twopence_connection twopence_conn_t;
-typedef struct twopence_connection_pool twopence_conn_pool_t;
-
-typedef const struct semantics twopence_conn_semantics_t;
-struct semantics {
-	bool		(*process_request)(twopence_transaction_t *, twopence_buf_t *);
-};
+#include "connection.h"
 
 #define DEFAULT_COMMAND_TIMEOUT	12	/* seconds */
 
-extern twopence_conn_t *	twopence_conn_new(twopence_conn_semantics_t *semantics, twopence_sock_t *client_sock, unsigned int client_id);
-extern void			twopence_conn_free(twopence_conn_t *conn);
-extern unsigned int		twopence_conn_fill_poll(twopence_conn_t *conn, twopence_pollinfo_t *pinfo);
-extern bool			twopence_conn_process_packet(twopence_conn_t *conn, twopence_buf_t *bp);
-extern bool			twopence_conn_process(twopence_conn_t *conn);
-
-extern twopence_conn_pool_t *	twopence_conn_pool_new(void);
-extern void			twopence_conn_pool_add_connection(twopence_conn_pool_t *pool, twopence_conn_t *conn);
-extern bool			twopence_conn_pool_poll(twopence_conn_pool_t *pool);
-
-extern void			server_run(twopence_sock_t *);
+extern void		server_run(twopence_sock_t *);
 
 #define AUDIT(fmt, args...) \
 	do { \
